+++ source/core/core.vcxproj.filters
@@ -69,6 +69,9 @@
     <ClInclude Include="slang-random-generator.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="slang-shared-library.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
     <ClInclude Include="slang-string-slice-pool.h">
       <Filter>Header Files</Filter>
     </ClInclude>
@@ -114,6 +117,9 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="slang-random-generator.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="slang-shared-library.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="slang-string-slice-pool.cpp">
