+++ source/core/slang-shared-library.cpp
