+++ source/core/slang-shared-library.h
